requirements.txt
torch
torchvision
numpy
pillow
gradio
gdown
